tried long int

+++ 0x07-pointers_arrays_strings/8-print_diagsums.c
@@ -11,8 +11,8 @@
 void print_diagsums(int *a, int size)
 {
 	int i = 0;
-	unsigned int sum1 = 0;
-	unsigned int sum2 = 0;
+	long int sum1 = 0;
+	long int sum2 = 0;
 
 	for (i = 0; i < size * size; i++)
 	{
